Add endpoint to get users grouped by team
Introduced a new HTTP POST endpoint `GetUsersGroupedByTeam` in the `UsersController` class. This endpoint is mapped to the route `"get-grouped-by-team"` and processes a `PaginationInputDto` from the request body. It returns a paginated list of users grouped by team, utilizing `_userService.GetUsersGroupedByTeamAsync` for data retrieval. Pagination metadata is included in the response headers.